third_party/blink/web_tests/virtual/offsetparent-old-behavior/fast/dom/shadow/offset-parent-does-not-leak-ua-shadow-expected.txt

offsetParent should not leak nodes in user agent Shadow DOM.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL child.offsetParent should be [object HTMLDivElement]. Was [object HTMLDivElement].
PASS child.offsetParent is container
PASS child.offsetParent is container
PASS child.offsetParent is container
PASS successfullyParsed is true

TEST COMPLETE
